Moved header data extraction to common lib
